getting closer to validation
I got a function to start checking through each input field to make sure the input isn't empty. The console is saving the values of the inputs now.

On the button click, it will run the function. However, I think I should be able to run some sort of forEach loop, to check if the input is blank. Then depending on the index of the input, display the corresponding errorIcon and span (visibility = visible).

I also can't get the password to save. If I change the type attribute to simply text, then once the button is submitted, do I change that attribute to password, so it's visually concealed?